# frigappapi/src/main/java/com/frigapp/frigappapi/controller/api/ProductsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.frigapp.frigappapi.controller.api;

import com.frigapp.frigappapi.model.CreateProductRequest;
import com.frigapp.frigappapi.model.ErrorResponse;
import com.frigapp.frigappapi.model.ProductsBarcodeBarcodeGet200Response;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.context.request.NativeWebRequest;

import java.util.Optional;
import java.util.UUID;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-26T22:36:13.511413300+01:00[Europe/Paris]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Products", description = "Référentiel produits (Open Food Facts)")
public interface ProductsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /products/barcode/{barcode} : Rechercher un produit par code-barres
     * Cherche dans la base locale, puis sur Open Food Facts si introuvable. Met le résultat en cache.
     *
     * @param barcode Code-barres EAN-13 (required)
     * @return Produit trouvé (status code 200)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Produit introuvable (status code 404)
     *         or Erreur Open Food Facts (status code 502)
     */
    @Operation(
        operationId = "productsBarcodeBarcodeGet",
        summary = "Rechercher un produit par code-barres",
        description = "Cherche dans la base locale, puis sur Open Food Facts si introuvable. Met le résultat en cache.",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Produit trouvé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductsBarcodeBarcodeGet200Response.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Produit introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "502", description = "Erreur Open Food Facts", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/barcode/{barcode}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ProductsBarcodeBarcodeGet200Response> productsBarcodeBarcodeGet(
        @Parameter(name = "barcode", description = "Code-barres EAN-13", required = true, in = ParameterIn.PATH) @PathVariable("barcode") String barcode
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"suggestedExpiryDays\" : 365, \"imageUrl\" : \"https://images.openfoodfacts.org/images/products/301/762/042/2003/front.jpg\", \"name\" : \"Nutella\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"source\" : \"open_food_facts\", \"category\" : \"Pâtes à tartiner\", \"barcode\" : \"3017620422003\", \"brand\" : \"Ferrero\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /products : Créer un produit manuellement
     * Crée un produit dans le référentiel local quand le scan ne donne aucun résultat.
     *
     * @param createProductRequest  (required)
     * @return Produit créé (status code 201)
     *         or Données invalides (status code 400)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Code-barres déjà existant (status code 409)
     */
    @Operation(
        operationId = "productsPost",
        summary = "Créer un produit manuellement",
        description = "Crée un produit dans le référentiel local quand le scan ne donne aucun résultat.",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Produit créé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductsBarcodeBarcodeGet200Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Données invalides", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Code-barres déjà existant", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<ProductsBarcodeBarcodeGet200Response> productsPost(
        @Parameter(name = "CreateProductRequest", description = "", required = true) @Valid @RequestBody CreateProductRequest createProductRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"data\" : { \"suggestedExpiryDays\" : 365, \"imageUrl\" : \"https://images.openfoodfacts.org/images/products/301/762/042/2003/front.jpg\", \"name\" : \"Nutella\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"source\" : \"open_food_facts\", \"category\" : \"Pâtes à tartiner\", \"barcode\" : \"3017620422003\", \"brand\" : \"Ferrero\" }, \"status\" : \"success\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    /**
     * DELETE /products/{productId} : Supprimer un produit manuel
     * Supprime un produit uniquement s&#39;il a été créé manuellement (source&#x3D;manual) et n&#39;est plus référencé par aucun item de frigo.
     *
     * @param productId  (required)
     * @return Produit supprimé (status code 204)
     *         or Token JWT absent ou invalide (status code 401)
     *         or Le produit n&#39;est pas de source manuelle (status code 403)
     *         or Ressource introuvable (status code 404)
     *         or Produit encore référencé par des items de frigo (status code 409)
     */
    @Operation(
            operationId = "productsProductIdDelete",
            summary = "Supprimer un produit manuel",
            description = "Supprime un produit uniquement s'il a été créé manuellement (source=manual) et n'est plus référencé par aucun item de frigo.",
            tags = { "Products" },
            responses = {
                    @ApiResponse(responseCode = "204", description = "Produit supprimé"),
                    @ApiResponse(responseCode = "401", description = "Token JWT absent ou invalide", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
                    }),
                    @ApiResponse(responseCode = "403", description = "Le produit n'est pas de source manuelle", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
                    }),
                    @ApiResponse(responseCode = "409", description = "Produit encore référencé par des items de frigo", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
                    })
            },
            security = {
                    @SecurityRequirement(name = "BearerAuth")
            }
    )
    @RequestMapping(
            method = RequestMethod.DELETE,
            value = "/products/{productId}",
            produces = { "application/json" }
    )

    default ResponseEntity<HttpStatus> productsProductIdDelete(
            @Parameter(name = "productId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("productId") UUID productId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"error\" : { \"code\" : \"FRIDGE_NOT_FOUND\", \"message\" : \"Le frigo demandé n'existe pas.\" }, \"status\" : \"error\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}