# frigappapi/src/main/resources/openapi/openapi_frigapp_v1.0.yaml
openapi: 3.0.3
info:
  title: Frig'App API
  description: |
    API REST de Frig'App — application mobile de gestion de stock de frigo collaboratif.
    
    ## Authentification
    Toutes les routes protégées nécessitent un token JWT passé dans le header `Authorization: Bearer <token>`.
    
    ## Temps réel
    Les modifications sur les items et les membres sont diffusées via WebSocket (STOMP over SockJS) sur `ws://api.frigapp.com/v1/ws`.
    
    ## Conventions
    - Toutes les réponses sont enveloppées dans `{ "status": "success"|"error", "data": ... }`
    - Les identifiants sont des UUID v4
    - Les timestamps sont en ISO 8601 UTC
  version: 1.0.0
  contact:
    name: Équipe Frig'App
servers:
  - url: https://api.frigapp.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Développement local (Spring Boot)
...
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── SCHEMAS ─────────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success]
          example: success
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        error:
          type: object
          properties:
            code:
              type: string
              example: FRIDGE_NOT_FOUND
            message:
              type: string
              example: Le frigo demandé n'existe pas.

    # ── USER ──────────────────────────────────────────────────────────────────
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        barcode:
          type: string
          example: "3017620422003"
        name:
          type: string
          example: Nutella
        brand:
          type: string
          nullable: true
          example: Ferrero
        category:
          type: string
          nullable: true
          example: Pâtes à tartiner
        imageUrl:
          type: string
          format: uri
          nullable: true
          example: https://images.openfoodfacts.org/images/products/301/762/042/2003/front.jpg
        source:
          type: string
          enum: [open_food_facts, manual]
          example: open_food_facts
        suggestedExpiryDays:
          type: integer
          nullable: true
          description: Durée de conservation suggérée en jours selon la catégorie
          example: 365
    CreateProductRequest:
      type: object
      required: [barcode, name]
      properties:
        barcode:
          type: string
          example: "1234567890123"
        name:
          type: string
          example: Mon produit
        brand:
          type: string
          nullable: true
        category:
          type: string
          nullable: true

    # ── FRIDGE ITEM ───────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Token JWT absent ou invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            error:
              code: UNAUTHORIZED
              message: Token JWT absent ou invalide.
...
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ── PATHS ─────────────────────────────────────────────────────────────────────
paths:
...
  /products/barcode/{barcode}:
    parameters:
      - name: barcode
        in: path
        required: true
        schema:
          type: string
        description: Code-barres EAN-13
        example: "3017620422003"
    get:
      tags: [Products]
      summary: Rechercher un produit par code-barres
      description: Cherche dans la base locale, puis sur Open Food Facts si introuvable. Met le résultat en cache.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Produit trouvé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Produit introuvable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: error
                error:
                  code: PRODUCT_NOT_FOUND
                  message: Produit introuvable localement et sur Open Food Facts.
        '502':
          description: Erreur Open Food Facts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products:
    post:
      tags: [Products]
      summary: Créer un produit manuellement
      description: Crée un produit dans le référentiel local quand le scan ne donne aucun résultat.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Produit créé
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Code-barres déjà existant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── ITEMS ─────────────────────────────────────────────────────────────────
...